Seperate Component Render & App Component for Shallow Enzyme testing + added basic enzyme tests
