+++ package.json
@@ -1,7 +1,8 @@
 {
   "name": "",
   "jest": {
-    "verbose": true
+    "verbose": true,
+    "setupTestFrameworkScriptFile": "./client/tests/test-env.js"
   },
   "version": "",
   "description": "",
@@ -32,6 +33,8 @@
   "devDependencies": {
     "babel-core": "^6.26.0",
     "babel-jest": "^22.4.1",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
     "eslint": "^4.18.1",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
@@ -39,6 +42,7 @@
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.7.0",
     "jest": "^22.4.2",
+    "jest-enzyme": "^4.2.0",
     "regenerator-runtime": "^0.11.1",
     "supertest": "^3.0.0"
   }
